+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * [Feature] Adds `ggplot` plotting API (histogram and boxplot)
 * [Fix] Clearer error when using `--with` with snippets that do not exist (#257)
 * [Fix] Pytds now automatically compatible
+* [Feature] Adds `%%config SqlMagic.polars_dataframe_kwargs = {...}`
+* [Fix] Jupysql with autopolars crashes when schema cannot be inferred from the first 100 rows ([#312](https://github.com/ploomber/jupysql/issues/312))
 ## 0.6.6 (2023-03-16)
 
 * [Fix] Pinning SQLAlchemy 1.x